requirements.txt
streamlit
google-generativeai
pandas
pypdf
pdfplumber
chromadb
langchain-google-genai